+++ src/nonim/backend/test/cases/expression_operator/expected.c
@@ -1,4 +1,4 @@
-static void operators (int const x, int const y) {
+static void operators (int64_t const x, int64_t const y) {
   (void)(x / y);
   (void)(x % y);
   (void)(x << y);
