+++ src/nonim/backend/test/cases/procedure_exported/expected.c
@@ -1,3 +1,3 @@
-int visible (int const x) {
+int64_t visible (int64_t const x) {
   return x;
 }
