+++ src/nonim/backend/test/cases/variable_exported/expected.c
@@ -1,1 +1,1 @@
-int const answer = 42;
+int64_t const answer = 42;
