+++ src/nonim/backend/test/cases/variable_var/expected.c
@@ -1,1 +1,1 @@
-static int counter = 0;
+static int64_t counter = 0;
